refactor: [BE] 유효한 메뉴와 메뉴 옵션을 만들어주는 함수, Entity layer로 분리
- 주문 비즈니스 로직을 처리하는 다른 함수들 처럼 도메인 영역으로 분리
- 또한 MenuOption을 DB로 부터 조회해야되는데 해당 부분은 함수 내부에서 DB를 접근한느 것이 아닌 인자로 받는 것으로 수정
